Use Slang instead of DXC in ShaderCooker
Make ShaderCooker use Slang instead of DXC
Put shader permutations in a .shaderpack file instead of individual files
Add support for loading shaders that are already in memory and owned by Game
Make Present function use Blit pipelines owned by Game using a callback
Add OnValueChanged callbacks to CVAR system
Do preparations for DescriptorSet refactors
Update SpirvReflect
Fix warning in TextureHandlerVK
Fix missing PopMarker in Present function
Add ShaderCooker support for outputting individual .spv files for debugging

+++ Source/Renderer/Renderer/Renderers/Vulkan/Backend/TextureHandlerVK.cpp
@@ -364,7 +364,7 @@ namespace Renderer
                     // For each mip level, calculate the reduced dimensions.
                     u32 mipWidth = Math::Max(1u, desc.width >> i);
                     u32 mipHeight = Math::Max(1u, desc.height >> i);
-                    mipUploadSize += mipWidth * mipHeight * FormatTexelSize(FormatConverterVK::ToVkFormat(texture->desc.format));
+                    mipUploadSize += static_cast<u64>(static_cast<f64>(mipWidth) * static_cast<f64>(mipHeight) * FormatTexelSize(FormatConverterVK::ToVkFormat(texture->desc.format)));
                 }
                 texture->uploadSize = mipUploadSize;
             }
